refactor tendances
