+++ src/main/Components/analyses/tendances/graphs/GraphAnalyseTendances.constant.ts
