+++ src/main/Models/analyses/tendances/AnalyseSoldesCategorie.model.ts
