+++ src/main/Models/analyses/tendances/AnalyseSoldesTimelineItem.model.ts
